Fixes to scadr scale.
